chore(tests): restructure test directory to mirror source architecture (#193)
Move test files into subdirectories matching py_modules/ layout
(services/, adapters/, domain/, models/, lib/). Merge
test_collections.py into test_library.py. Fix test isolation bugs
where tests mutated global decky mock state instead of injecting
PersistenceAdapter directly.
