+++ README.md
@@ -102,7 +102,3 @@ This plugin stands on the shoulders of some great projects:
 ## License
 
 GPL-3.0
-
-## Star History
-
-[![Star History Chart](https://api.star-history.com/svg?repos=danielcopper/decky-romm-sync&type=Date)](https://star-history.com/#danielcopper/decky-romm-sync&Date)
