+++ pytest.ini
@@ -1,3 +1,4 @@
 [pytest]
 asyncio_mode = auto
 testpaths = tests
+addopts = --import-mode=importlib
