+++ tests/adapters/romm/test_api_router.py
