+++ tests/adapters/romm/test_api_v47.py
