+++ tests/adapters/test_persistence.py
