+++ tests/adapters/test_steam_config.py
