+++ tests/adapters/test_steamgriddb.py
