+++ tests/domain/test_emulator_tag.py
