+++ tests/domain/test_rom_files.py
