+++ tests/domain/test_shortcut_data.py
