+++ tests/domain/test_steam_categories.py
