+++ tests/lib/test_errors.py
