+++ tests/services/test_artwork.py
