+++ tests/services/test_firmware.py
