+++ tests/services/test_rom_removal.py
